fix(ci): add commit hash to cache keys
AzureDevOps caches are immutable. This means they cannot be updated if the key
doesn't change across commits.
Instead, we add a commit hash to the key, making it distinct for each commit
however the cache can still fallback to the restoreKey, using the lastest
cache with partially matched key prefix.